Move company types to types dir
