+++ app/types/company.ts
